move new parsing logic to the main dogstatsd package
