+++ pkg/dogstatsd/convert.go
@@ -1,4 +1,4 @@
-package dogstatsd_tmp
+package dogstatsd
 
 import (
 	"bytes"
@@ -19,6 +19,30 @@ var (
 
 	getTags tagRetriever = tagger.Tag
 )
+
+func parseMetricMessage(message []byte, namespace string, namespaceBlacklist []string, defaultHostname string) (*metrics.MetricSample, error) {
+	sample, err := parseMetricSample(message)
+	if err != nil {
+		return nil, err
+	}
+	return convertMetricSample(sample, namespace, namespaceBlacklist, defaultHostname), nil
+}
+
+func parseEventMessage(message []byte, defaultHostname string) (*metrics.Event, error) {
+	sample, err := parseEvent(message)
+	if err != nil {
+		return nil, err
+	}
+	return convertEvent(sample, defaultHostname), nil
+}
+
+func parseServiceCheckMessage(message []byte, defaultHostname string) (*metrics.ServiceCheck, error) {
+	sample, err := parseServiceCheck(message)
+	if err != nil {
+		return nil, err
+	}
+	return convertServiceCheck(sample, defaultHostname), nil
+}
 
 func convertTags(tags [][]byte, defaultHostname string) ([]string, string) {
 	if len(tags) == 0 {
