+++ pkg/dogstatsd/parse_events_test.go
@@ -1,4 +1,4 @@
-package dogstatsd_tmp
+package dogstatsd
 
 import (
 	"testing"
